Summarize safety data

This function creates the text and summary table for the report.

asr_safety_summary(data, period_data, trial_type, n_pat_e)

Arguments

data

SAE data

period_data

SAE data restricted to a specific period

trial_type

trial type (imp, medical device or other)

n_pat_e

Number of enrolled participants

Value

txt

The sentences required for the report

tab

The summary table required for the report

tab_map

Mapping between the variable name in tab and a nicer label

Examples

data(asr_sae)
# IMP
prepped <- asr_dataprep(asr_sae,
                        period_from = as.Date("2020-10-10"),
                        period_to = as.Date("2021-10-10"))
asr_safety_summary(data = prepped$data, period_data = prepped$period_data,
                   "imp", 60)
#> $txt
#> [1] "During the reporting period, 20 of 60 participants (33.3 %) reported a total of 20 serious adverse events (SAEs)."                                                        
#> [2] "5 of 20 SAEs (25.0 %) were classified 'related' to the IMP. The most frequent related SAEs documented were poor sleep (2), spontaneous unconsciousness (2), headache (1)."
#> [3] "6 Suspected Unexpected Serious Adverse Reactions (SUSARs) occurred during the reporting period, which have been notified to the Swiss competent authorities."             
#> 
#> $tab
#> # A tibble: 2 × 5
#>   desc                                                   fatal   sae  sadr susar
#>   <chr>                                                  <int> <int> <int> <int>
#> 1 Number of cases (during reporting period)                  0    20     5     6
#> 2 Number of cases (cumulative) since the start of the c…     3    20     5     6
#> 
#> $tab_map
#>   col_key
#> 1    desc
#> 2   fatal
#> 3     sae
#> 4    sadr
#> 5   susar
#>                                                                     name
#> 1                                                                       
#> 2                                                            Fatal cases
#> 3                                           Serious Adverse Events, SAEs
#> 4                  Serious Adverse Drug Reactions, SADRs (only for IMPs)
#> 5 Suspected Unexpected Serious Adverse Reactions, SUSARs (only for IMPs)
#> 

# medical devices
prepped <- asr_dataprep(asr_sae, period_from = as.Date("2020-10-10"),
                        period_to = as.Date("2021-10-10"), trial_type = "m")
summ <- asr_safety_summary(data = prepped$data,
                           period_data = prepped$period_data, "m", 60)

# other trial
asr_safety_summary(data = prepped$data, period_data = prepped$period_data, "o", 60)
#> $txt
#> [1] "During the reporting period, 20 of 60 participants (33.3 %) reported a total of 20 serious adverse events (SAEs; with possible relationship to the intervention)."
#> [2] "The most frequent SAEs documented were spontaneous unconsciousness (7), headache (5), poor sleep (4)."                                                            
#> 
#> $tab
#> # A tibble: 2 × 3
#>   desc                                                               fatal   sae
#>   <chr>                                                              <int> <int>
#> 1 Number of cases (during reporting period)                              0    20
#> 2 Number of cases (cumulative) since the start of the clinical trial     3    20
#> 
#> $tab_map
#>   col_key                         name
#> 1    desc                             
#> 2   fatal                  Fatal cases
#> 3     sae Serious Adverse Events, SAEs
#>